feat: Add jest for unit and e2e tests

+++ src/use-cases/user/services/delete-user.service.ts
@@ -8,7 +8,12 @@ export class DeleteUserService {
         private readonly userRepository: UserRepository
     ) {}
 
-    async execute(id: number): Promise<void> {
+    async execute(id: number): Promise<boolean> {
+        const user = await this.userRepository.getById(id);
+        if (!user) {
+            return false;
+        }
         await this.userRepository.delete(id);
+        return true;
     }
 }
